+++ packages/dropdown_button2/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Re-lay out dropdown menu on ancestor scroll.
 - Add `barrierBlocksInteraction` to allow interaction with underlying widgets while the dropdown menu is open.
 - Properly dispose internal FocusNode when replaced by an external FocusNode.
+- Add `mouseCursor` to the dropdown button and MenuItemStyleData to customize the mouse cursor when hovering, closes #416.
 
 ## 3.0.0